page_appbar chatroom 추가
